Add a catalog of documents (#1)
* Add a catalog of documents

Pull document details from metadata files in buckets specified in the config.
Get metadata files for each bucket, parse into documents. Use the prefix option
to narrow down the files we are searching over. All metadata files that Access
is uploading start with "MADA".

Store the document catalog in an ets table and provide a lookup function. This
allows searching documents by the last name, first name, and date of birth of
the people the documents are for.

Add dialyxir as a dependency.