Add a document search page
Validate that at least one field is entered.
Parse and validate the date of birth.

Adapt Catalog.lookup to take a Query struct as input.

Remove the unused page controller, view, and template.

Add ecto and phoenix_ecto as dependencies.

The ability to click a search result and view the document will come in
a later PR.