Migrates document viewer to Keycloak
In order to do this we needed to add the ueberauth_oidc library which
then involved updating a few other deps including Phoenix. But this
should now enable a user with the correct role logging into the app via
keycloak. It will reject any user who does not have the correct role.